services/budget-service/pyproject.toml
[project]
name = "finans-tracker-budget-service"
version = "0.1.0"
description = "Budget microservice for Finans Tracker"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27.0",
]

[dependency-groups]
dev = [
    "ruff",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "respx>=0.21.0",
    "testcontainers[postgres]>=4.14.2",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
